most decorators and error handling works, working on tags

+++ backend/app/routes/comment.py
@@ -7,23 +7,24 @@ from ..models.db import db
 from ..utils.errors import ValidationError
 from ..utils.decorator import (
     login_check,
-    question_exist_check,
-    answer_exist_check,
-    comment_for_question_exist_check,
-    comment_for_question_ownership_check,
-    comment_for_answer_exist_check,
-    comment_for_answer_ownership_check,
+    # question_exist_check,
+    # answer_exist_check,
+    # comment_for_question_exist_check,
+    # comment_for_question_ownership_check,
+    # comment_for_answer_exist_check,
+    # comment_for_answer_ownership_check,
     collect_query_params,
-    existence_check,authorization_check,comment_owner_check
+    existence_check,authorization_check,owner_check
 )
 
 bp = Blueprint("comment", __name__, url_prefix="/api/questions")
 
 
 @bp.route("/<int:question_id>/comments", methods=["GET"])
-@question_exist_check
+# @question_exist_check
+@existence_check(("Question","question_id"))
 @collect_query_params(Comment)
-def get_all_comments_for_question(question_id, page, per_page, sort_column, sort_order):
+def get_all_comments_for_question(question_id, question, page, per_page, sort_column, sort_order):
     comments = (
         Comment.query.filter_by(content_id=question_id, content_type="question")
         .order_by(sort_order(sort_column))
@@ -42,10 +43,11 @@ def get_all_comments_for_question(question_id, page, per_page, sort_column, sort
 
 
 @bp.route("/<int:question_id>/answers/<int:answer_id>/comments", methods=["GET"])
-@question_exist_check
+# @question_exist_check
+@existence_check(("Question","question_id"))
 @collect_query_params(Comment)
 def get_all_comments_for_an_answer(
-    question_id, answer_id, page, per_page, sort_column, sort_order
+    question_id,question, answer_id, page, per_page, sort_column, sort_order
 ):
     comments = (
         Comment.query.filter_by(content_id=answer_id, content_type="answer")
@@ -64,8 +66,9 @@ def get_all_comments_for_an_answer(
 
 
 @bp.route("/<int:question_id>/allcomments", methods=["GET"])
-@question_exist_check
-def get_all_comments(question_id):
+# @question_exist_check
+@existence_check(("Question","question_id"))
+def get_all_comments(question_id,question):
     comments_list = []
 
     question_comments = Comment.query.filter_by(
@@ -96,8 +99,11 @@ def get_all_comments(question_id):
 def create_comment_for_question(question_id,question):
     data = request.get_json()
     content = data.get("content")
+    errors = []
     if not content:
-        return jsonify({"error": "content is required"}), 400
+        errors.append(("content","Data is required"))
+    if errors:
+        raise ValidationError(errors=errors)
     new_comment = Comment(
         user_id=current_user.id,
         content=content,
@@ -115,14 +121,20 @@ def create_comment_for_question(question_id,question):
 # @comment_for_question_exist_check
 @existence_check(("Question","question_id"),("Comment","comment_id"))
 # @comment_for_question_ownership_check
-@authorization_check(comment_owner_check,"comment")
+@authorization_check(owner_check,"comment")
 def edit_comment_for_question(question_id,question,comment_id,comment):
     data = request.get_json()
-    new_content = data.get("content")
-    if not new_content:
-        return jsonify({"error": "content is required"}), 400
+    content = data.get("content")
+    # print(comment.content_id)
+    errors = []
+    if not content:
+        errors.append(("content","Data is required"))
+    elif comment.content_id != question.id:
+        errors.append(("comment","comment does not belong to this question"))
+    if errors:
+        raise ValidationError(errors=errors)
     # comment = Comment.query.get(comment_id)
-    comment.content = new_content
+    comment.content = content
     db.session.commit()
     return jsonify({"comment": comment.to_dict()}), 200
 
@@ -130,10 +142,17 @@ def edit_comment_for_question(question_id,question,comment_id,comment):
 @bp.route("/<int:question_id>/comments/<int:comment_id>", methods=["DELETE"])
 # @csrf_protect
 @login_check
-@comment_for_question_exist_check
-@comment_for_question_ownership_check
-def delete_comment_for_question(question_id, comment_id):
-    comment = Comment.query.get(comment_id)
+# @comment_for_question_exist_check
+# @comment_for_question_ownership_check
+@existence_check(("Question","question_id"),("Comment","comment_id"))
+@authorization_check(owner_check,"comment")
+def delete_comment_for_question(question_id,question,comment_id,comment):
+    # comment = Comment.query.get(comment_id)
+    errors = []
+    if comment.content_id != question.id:
+        errors.append(("comment","comment does not belong to this question"))
+    if errors:
+        raise ValidationError(errors=errors)
     db.session.delete(comment)
     db.session.commit()
     return jsonify({"message": "comment for question deleted"}), 200
@@ -142,13 +161,21 @@ def delete_comment_for_question(question_id, comment_id):
 @bp.route("/<int:question_id>/answers/<int:answer_id>/comments", methods=["POST"])
 # @csrf_protect
 @login_check
-@question_exist_check
-@answer_exist_check
-def create_comment_for_answer(question_id, answer_id):
+# @question_exist_check
+# @answer_exist_check
+@existence_check(("Question","question_id"),("Answer","answer_id"))
+def create_comment_for_answer(question_id,question,answer_id,answer):
     data = request.get_json()
     content = data.get("content")
+    # if not content:
+    #     return jsonify({"error": "content is required"}), 400
+    errors = []
+    if answer.question_id != question.id:
+        errors.append(("answer","answer does not belong to this question"))
     if not content:
-        return jsonify({"error": "content is required"}), 400
+        errors.append(("comment","Data is required"))
+    if errors:
+        raise ValidationError(errors=errors)
     new_comment = Comment(
         user_id=current_user.id,
         content=content,
@@ -166,15 +193,23 @@ def create_comment_for_answer(question_id, answer_id):
 )
 # @csrf_protect
 @login_check
-@comment_for_answer_exist_check
-@comment_for_answer_ownership_check
-def edit_comment_for_answer(question_id, answer_id, comment_id):
+# @comment_for_answer_exist_check
+# @comment_for_answer_ownership_check
+@existence_check(("Question","question_id"),("Answer","answer_id"),("Comment","comment_id"))
+@authorization_check(owner_check,"comment")
+def edit_comment_for_answer(question_id, question, answer_id, answer, comment_id, comment):
     data = request.get_json()
-    new_content = data.get("content")
-    if not new_content:
-        return jsonify({"error": "content is required"}), 400
-    comment = Comment.query.get(comment_id)
-    comment.content = new_content
+    content = data.get("content")
+    errors = []
+    if answer.question_id != question.id:
+        errors.append(("answer","answer does not belong to this question"))
+    if comment.content_id != answer.id:
+        errors.append(("comment","comment does not belong to this answer"))
+    if not content:
+        errors.append(("comment","Data is required"))
+    if errors:
+        raise ValidationError(errors=errors)
+    comment.content = content
     db.session.commit()
     return jsonify({"comment": comment.to_dict()}), 200
 
@@ -185,10 +220,19 @@ def edit_comment_for_answer(question_id, answer_id, comment_id):
 )
 # @csrf_protect
 @login_check
-@comment_for_answer_exist_check
-@comment_for_answer_ownership_check
-def delete_comment_for_answer(question_id, answer_id, comment_id):
-    comment = Comment.query.get(comment_id)
+# @comment_for_answer_exist_check
+# @comment_for_answer_ownership_check
+@existence_check(("Question","question_id"),("Answer","answer_id"),("Comment","comment_id"))
+@authorization_check(owner_check,"comment")
+def delete_comment_for_answer(question_id, question, answer_id, answer, comment_id, comment):
+    # comment = Comment.query.get(comment_id)
+    errors = []
+    if answer.question_id != question.id:
+        errors.append(("answer","answer does not belong to this question"))
+    if comment.content_id != answer.id:
+        errors.append(("comment","comment does not belong to this answer"))
+    if errors:
+        raise ValidationError(errors=errors)
     db.session.delete(comment)
     db.session.commit()
     return jsonify({"message": "comment for answer deleted"}), 200
